+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.1.0
 - [Add layers attribute](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/7)
+- [Add secret environment variables attribute](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/8)
 
 
 ## v1.0.0